exercise 4.19: blog list expansions, step7

+++ part4/controllers/blogs.js
@@ -1,4 +1,5 @@
 const router = require('express').Router();
+const jwt = require('jsonwebtoken');
 require('express-async-errors');
 const Blog = require('../models/blog');
 const User = require('../models/user');
@@ -9,25 +10,29 @@ router.get('/', async (request, response) => {
 });
 
 router.post('/', async (request, response) => {
-    const user = await User.findById(request.body.user);
-    if (!user) {
-        return response.status(400).json({ error: 'unknown user' });
+    const authenticatedUser = await getAuthenticatedUser(request);
+    if (!authenticatedUser) {
+        return response.status(401).json({ error: 'token missing or invalid' });
     }
-    delete request.body.user;
 
     const blog = new Blog({
         ...request.body,
-        user: user._id,
+        user: authenticatedUser._id,
     });
     const result = await blog.save();
 
-    user.blogs.push(result._id);
-    await user.save();
+    authenticatedUser.blogs.push(result._id);
+    await authenticatedUser.save();
 
     response.status(201).json(result);
 });
 
 router.delete('/:id', async (request, response) => {
+    const authenticatedUser = await getAuthenticatedUser(request);
+    if (!authenticatedUser) {
+        return response.status(401).json({ error: 'token missing or invalid' });
+    }
+
     const blog = await Blog.findByIdAndDelete(request.params.id);
     if (blog) {
         response.json(blog);
@@ -37,6 +42,11 @@ router.delete('/:id', async (request, response) => {
 });
 
 router.patch('/:id', async (request, response) => {
+    const authenticatedUser = await getAuthenticatedUser(request);
+    if (!authenticatedUser) {
+        return response.status(401).json({ error: 'token missing or invalid' });
+    }
+
     const blog = await Blog.findById(request.params.id);
     if (!blog) {
         return response.status(404).end();
@@ -46,5 +56,24 @@ router.patch('/:id', async (request, response) => {
     const savedBlog = await blog.save();
     response.json(savedBlog);
 });
+
+const getAuthenticatedUser = async request => {
+    const authorization = request.get('authorization');
+    if (!authorization?.toLowerCase().startsWith('bearer ')) {
+        return undefined;
+    }
+
+    const token = authorization.substring(7);
+    try {
+        const decodedToken = jwt.verify(token, process.env.SECRET);
+
+        return await User.findById(decodedToken.id);
+    } catch (e) {
+        if (e.name === 'JsonWebTokenError') {
+            return undefined;
+        }
+        throw e;
+    }
+};
 
 module.exports = router;
